# static/swagger-ui/swagger/openapi.yaml
swagger: "2.0"
info:
  title: JeeekAPI
  description: エンジニアのための自己実現SNSアプリ
  contact:
    name: tonouchi510
    email: tonouchi27@gmail.com
    url: https://github.com/tonouchi510/Jeeek/issues
  version: "0.1"
host: localhost:8080
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /v1/users:
    get:
      tags:
      - User
      summary: List user User
      description: ユーザーの一覧を返します。
      operationId: User#List user
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserJeeekUserResponseCollection'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/UserListUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    put:
      tags:
      - User
      summary: Update user User
      description: 現在のセッションに紐付くユーザー情報を更新します。
      operationId: User#Update user
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      - name: Update UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserUpdateUserRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserUpdateUserResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/UserUpdateUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    delete:
      tags:
      - User
      summary: Delete user User
      description: 現在のセッションに紐づくユーザーを削除します。
      operationId: User#Delete user
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "204":
          description: No Content response.
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/UserDeleteUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /v1/users/{user_id}:
    get:
      tags:
      - User
      summary: Get user User
      description: 指定したIDのユーザーの情報を返します。
      operationId: User#Get user
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: user_id
        in: path
        description: User id of firebase
        required: true
        type: string
        maxLength: 28
        minLength: 28
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserGetUserResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/UserGetUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /v1/users/me:
    get:
      tags:
      - User
      summary: Get current user User
      description: 現在のセッションに紐付くユーザの情報を返します。
      operationId: User#Get current user
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserGetCurrentUserResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/UserGetCurrentUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
definitions:
...
  JeeekUserResponse:
    title: 'Mediatype identifier: application/vnd.jeeek.user+json; view=default'
    type: object
    properties:
      email_address:
        type: string
        description: ユーザーのプライマリ メールアドレス
        example: keisuke.honda+testuser@ynu.jp
        format: email
      email_verified:
        type: boolean
        description: ユーザーのプライマリ メールアドレスが確認されているかどうか
        example: false
      phone_number:
        type: string
        description: ユーザのメインの電話番号
        example: "08079469367"
        pattern: ^\+?[\d]{10,}$
      photo_url:
        type: string
        description: ユーザーの写真 URL
        example: https://imageurl.com
      user_id:
        type: string
        description: User id of firebase
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      user_name:
        type: string
        description: ユーザーの表示名
        example: keisuke.honda
        minLength: 1
        maxLength: 20
    description: user response (default view)
    example:
      email_address: keisuke.honda+testuser@ynu.jp
      email_verified: true
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    required:
    - user_id
    - user_name
    - email_address
    - photo_url
    - phone_number
    - email_verified
  UserDeleteUserUnauthorizedResponseBody:
    title: UserDeleteUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: A architecto.
  UserGetCurrentUserResponseBody:
    title: 'Mediatype identifier: application/vnd.jeeek.user+json; view=default'
    type: object
    properties:
      email_address:
        type: string
        description: ユーザーのプライマリ メールアドレス
        example: keisuke.honda+testuser@ynu.jp
        format: email
      email_verified:
        type: boolean
        description: ユーザーのプライマリ メールアドレスが確認されているかどうか
        example: false
      phone_number:
        type: string
        description: ユーザのメインの電話番号
        example: "08079469367"
        pattern: ^\+?[\d]{10,}$
      photo_url:
        type: string
        description: ユーザーの写真 URL
        example: https://imageurl.com
      user_id:
        type: string
        description: User id of firebase
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      user_name:
        type: string
        description: ユーザーの表示名
        example: keisuke.honda
        minLength: 1
        maxLength: 20
    description: Get Current UserResponseBody result type (default view)
    example:
      email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    required:
    - user_id
    - user_name
    - email_address
    - photo_url
    - phone_number
    - email_verified
  UserGetCurrentUserUnauthorizedResponseBody:
    title: UserGetCurrentUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Qui vel ut.
  UserGetUserResponseBody:
    title: 'Mediatype identifier: application/vnd.jeeek.user+json; view=default'
    type: object
    properties:
      email_address:
        type: string
        description: ユーザーのプライマリ メールアドレス
        example: keisuke.honda+testuser@ynu.jp
        format: email
      email_verified:
        type: boolean
        description: ユーザーのプライマリ メールアドレスが確認されているかどうか
        example: true
      phone_number:
        type: string
        description: ユーザのメインの電話番号
        example: "08079469367"
        pattern: ^\+?[\d]{10,}$
      photo_url:
        type: string
        description: ユーザーの写真 URL
        example: https://imageurl.com
      user_id:
        type: string
        description: User id of firebase
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      user_name:
        type: string
        description: ユーザーの表示名
        example: keisuke.honda
        minLength: 1
        maxLength: 20
    description: Get UserResponseBody result type (default view)
    example:
      email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    required:
    - user_id
    - user_name
    - email_address
    - photo_url
    - phone_number
    - email_verified
  UserGetUserUnauthorizedResponseBody:
    title: UserGetUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Et voluptas voluptas assumenda ut est aliquam.
  UserJeeekUserResponseCollection:
    title: 'Mediatype identifier: application/vnd.jeeek.user+json; type=collection;
      view=default'
    type: array
    items:
      $ref: '#/definitions/JeeekUserResponse'
    description: List UserResponseBody is the result type for an array of JeeekUserResponse
      (default view)
    example:
    - email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    - email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    - email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    - email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
  UserListUserUnauthorizedResponseBody:
    title: UserListUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Enim consectetur a qui quia dicta.
  UserUpdateUserRequestBody:
    title: UserUpdateUserRequestBody
    type: object
    properties:
      email_address:
        type: string
        description: ユーザーのプライマリ メールアドレス
        example: keisuke.honda+testuser@ynu.jp
        format: email
      phone_number:
        type: string
        description: ユーザのメインの電話番号
        example: "08079469367"
        pattern: ^\+?[\d]{10,}$
      photo_url:
        type: string
        description: ユーザーの写真 URL
        example: https://imageurl.com
      user_name:
        type: string
        description: ユーザーの表示名
        example: keisuke.honda
        minLength: 1
        maxLength: 20
    example:
      email_address: keisuke.honda+testuser@ynu.jp
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_name: keisuke.honda
  UserUpdateUserResponseBody:
    title: 'Mediatype identifier: application/vnd.jeeek.user+json; view=default'
    type: object
    properties:
      email_address:
        type: string
        description: ユーザーのプライマリ メールアドレス
        example: keisuke.honda+testuser@ynu.jp
        format: email
      email_verified:
        type: boolean
        description: ユーザーのプライマリ メールアドレスが確認されているかどうか
        example: true
      phone_number:
        type: string
        description: ユーザのメインの電話番号
        example: "08079469367"
        pattern: ^\+?[\d]{10,}$
      photo_url:
        type: string
        description: ユーザーの写真 URL
        example: https://imageurl.com
      user_id:
        type: string
        description: User id of firebase
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      user_name:
        type: string
        description: ユーザーの表示名
        example: keisuke.honda
        minLength: 1
        maxLength: 20
    description: Update UserResponseBody result type (default view)
    example:
      email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    required:
    - user_id
    - user_name
    - email_address
    - photo_url
    - phone_number
    - email_verified
  UserUpdateUserUnauthorizedResponseBody:
    title: UserUpdateUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Eaque molestiae omnis eaque consequatur rerum cum.
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: Secures endpoint by requiring a valid JWT token retrieved via the
      firebase.
    name: Authorization
    in: header
...